Factor out some support code
Patches such as #112 could use them.
